requirements.txt
streamlit
openai
langchain
langchain-community
python-dotenv
pypdf
python-docx
tiktoken
faiss-cpu
